minor fixes for import errors and changing keyword formatting in unit tests
