+++ nems/utils.py
@@ -83,3 +83,4 @@ def find_module(name, modelspec, find_all_matches=False):
                   target_module, target_i)
 
     return target_i
+
