+++ tests/test_load_recording.py
@@ -15,10 +15,9 @@ import shutil
 
 from nems.signal import load_signal_from_streams
 import nems.recording as recording
-from nems.uri import get_demo_recordings
 
 def test_load_recording():
-    get_demo_recordings()
+    recording.get_demo_recordings()
 
     nems_dir = os.path.abspath(os.path.dirname(__file__) + '/..')
     targz = nems_dir + '/recordings/TAR010c-18-1.tgz'
